requirements.txt
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-nmap==0.6.1
SQLAlchemy==1.2.8
Werkzeug==0.14.1
